Make starting integration time a configurable property.
The starting integration time was hardcoded to -0.25ns.  Hits earlier
than that time were removed.

However, if vertex smearing is enabled in time, then we may see legitimate
hits earlier than that (for FCC, the smearing width is about 1.9 ns).
Add a property to set this value, rather then hardcoding it.